B-3: past_due segura o acesso por 15 dias, e depois reduz
Regra do dono (2026-09-04), na varredura:
  ativa/trialing mais longa que o grant   → repara (mantido)
  past_due dentro de PAST_DUE_CARENCIA_DIAS → NÃO reduz, segura o acesso, LOGA
                                              e não alerta (dunning é estado
                                              esperado; alerta diário é ruído)
  past_due além do teto                  → REDUZ, mesmo com a assinatura viva
  sem assinatura ativa                   → reduz (9e, mantido)
  Stripe indisponível                    → não escreve, alerta (mantido)

O teto é o que mata a cauda infinita: `_find_active_subscription` considera
`past_due` viva e o `current_period_end` dela é o período JÁ vencido, então sem
o teto nunca haveria o que reparar e a conta ficaria pendurada para sempre.

`PAST_DUE_CARENCIA_DIAS = 15`, constante nomeada, sem env — decisão de produto,
não número técnico. O comentário registra o fato que o coordenador levantou: o
dunning padrão do Stripe retenta por volta de DUAS SEMANAS, então 15 fica na
BORDA; se as tentativas da conta forem mais longas, o teto corta alguém que o
Stripe ainda recuperaria. O caminho de volta é automático (o `invoice.paid` da
recuperação regrava o grant), então o pior caso é INTERMITÊNCIA, não perda
permanente.

`_observar` ganhou `alerta=None` para logar sem alertar.

Testes: dia 14 mantém, dia 16 reduz. Negativos rodados: carência 0 derruba o
dia 14; carência 9999 derruba o dia 16; remover o ramo derruba os dois. O
positivo do 9e continua vermelho sob "nunca reduz", junto com o caso 7.

Os casos de redução saíram para `tests/test_billing_grants_reducao.py` — o de
materialização passava de 350 de novo. Materialização é o grant ENTRAR; redução
é o grant SAIR. Cinco mutações refeitas depois do corte.
